fix http_response bugs

+++ http.h
@@ -62,7 +62,7 @@ struct set_logging
 int request(char *buf, struct http_request *request_info, struct set_logging *logging_info);	/* deal with http request */
 void clean_request(struct http_request *request_info);
 
-int response(struct http_response *response_info, char *resp_buf, int capacity, int *size);    /* deal with http response */
+int response(struct http_response *response_info, char *resp_buf, size_t capacity, size_t *size);    /* deal with http response */
 
 /* logging will return the length function written
  * return 0 if error
